Added normalisation of dependent variable

+++ AkkaNetNeuralNet/AkkaNetNeuralNet.Core/Model/IDependentExpectancyVariable.cs
@@ -2,6 +2,6 @@ namespace AkkaNetNeuralNet.Core.Model
 {
     public interface IDependentExpectancyVariable
     {
-        int AgeAtDeath { get; }
+        decimal AgeAtDeath { get; }
     }
 }
